backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0
supabase==2.11.0
google-generativeai==0.8.4
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.159.0
pdfplumber==0.11.4
pandas==2.2.3
httpx==0.28.1
python-multipart==0.0.20
cryptography==44.0.0
